attempt to fix build
this is caused becaused the hardhat task expects typechain but the typechain
can't run because the hardhat config imports the task

also described here: https://github.com/dethcrypto/TypeChain/issues/371
following a solution as in: https://github.com/mstable/merkle-drop/commit/c0e90331a2d6324e4b43364b530b6a2449a5c291
